requirements.txt
Flask
Flask-Admin
Flask-SQLAlchemy
Flask-Security>=1.7.5
Flask-SocketIO
Flask-Mqtt
Flask-JSglue
Eventlet
Gunicorn

